% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpselect.R
\name{grpselect}
\alias{grpselect}
\title{grpselect}
\usage{
grpselect(fgrp, x, y, B = 50, parallel = TRUE)
}
\arguments{
\item{fgrp}{the variable group object from `deCorr``}

\item{x}{the predictor matrix}

\item{y}{a dataframe of time to event and event status. The primary outcome status is coded 1, the secondary outcome as 2, etc. The censored is coded as 0.}

\item{B}{the number of bootstraps}

\item{parallel}{whether to use multiple cores for parallel computing. Default is TRUE.}
}
\value{
a list of 
\itemize{
\item{fselect:}{ Names of the selected variables.}
\item{prob:}{ the generalized ridge variable importance.}
\item{weight:}{ the inverse of the ridge variable importance.}
}
}
\description{
This is the split step, where variable in subgroups are selected
}
