% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossTransform.R
\name{lossTrans}
\alias{lossTrans}
\alias{mod_lsa}
\title{Linear Approximation of the object function}
\usage{
mod_lsa(obj, n)
}
\arguments{
\item{obj}{the regression object from R output}

\item{n}{the sample size}
}
\description{
Linear Approximation of the object function
}
