% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2select.R
\name{r2select}
\alias{r2select}
\title{Variable Selection with the candidate pool}
\usage{
r2select(x.tr, y.tr, fr, B, weight, prob, parallel = TRUE)
}
\arguments{
\item{x.tr}{the predictor matrix}

\item{y.tr}{the time and status object for survival}

\item{B}{times of bootstrap}

\item{weight}{variable weight}

\item{prob}{variable selection probability}
}
\description{
Perform variable selection with pooled candidates
}
