% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAEN-Package.R
\docType{package}
\name{RAEN-Package}
\alias{RAEN-Package}
\title{The Robust and Generalized Ensemble Approach for Variable Selection in High Dimensions}
\description{
We provide a novel solution to the variable selection problem in the ultra-high dimensional setting with a robust and generalized method
}
\author{
Han Sun and Xiaofeng Wang
}
