% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAEN.R
\name{RAEN}
\alias{RAEN}
\alias{predict.RAEN}
\title{Random Ensemble Variable Selection for Ultra-high Dimensional Data}
\usage{
RAEN(x, y, B, ngrp = floor(15 * ncol(x)/nrow(x)), parallel = TRUE, ncore = 2)

\method{predict}{RAEN}(obj, newdata, ...)
}
\arguments{
\item{x}{the predictor matrix}

\item{y}{the time and status object for survival}

\item{B}{times of bootstrap}

\item{ngrp}{the number of blocks to separate variables into. Default is 15*p/N, where p is the number of predictors and N is the sample size.}

\item{parallel}{Logical TRUE or FALSE. Whether to use multithread computing, which can save consideratable amount of time for high dimensional data}

\item{ncore}{Number of cores used for parallel computing, if parallel=TRUE}

\item{obj}{the RAEN object containing the variable selection results}

\item{newdata}{the predictor matrix for prediction}
}
\value{
a dataframe with the variable names and the regression coefficients

the linear predictor of the outcome risk
}
\description{
Perform variable selection for ultra-high dimensional data
}
