% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Toy data for demonstration}
\format{
A dataframe of 200 rows and 1002 columns.
}
\description{
This simulated datasets contains 1000 predictors, of which X1-X20, X41-X60 are true predictors.\\  The first two columns are time to event competing risks and status.
}
\keyword{datasets}
